ДЗ 9: добавлен generic на уровень класса и выделен интерфейс класса DbExecutorOrm, а сам класс переименован в DbExecutorOrmImpl. Изменены тесты с учетом изменений.
