+++ hw09-orm/src/test/java/ru/otus/lesson17/DbExecutorOrmTest.java
@@ -3,6 +3,7 @@ package ru.otus.lesson17;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
 import ru.otus.lesson17.orm.DbExecutorOrm;
+import ru.otus.lesson17.orm.DbExecutorOrmImpl;
 import ru.otus.lesson17.model.Account;
 import ru.otus.lesson17.model.User;
 
@@ -33,42 +34,43 @@ public class DbExecutorOrmTest {
             connection.setAutoCommit(false);
             createTables(connection);
 
-            DbExecutorOrm dbExecutorOrm = new DbExecutorOrm();
+            DbExecutorOrm<User> userOrm = new DbExecutorOrmImpl<>();
 
             // User
             User user = new User(null, "test", 5);
-            Long id = (Long) dbExecutorOrm.create(connection, user);
+            Long id = (Long) userOrm.create(connection, user);
             Assert.assertEquals(Long.valueOf(1L), id);
 
             user.setAge(25);
-            dbExecutorOrm.update(connection, user);
+            userOrm.update(connection, user);
 
             connection.commit();
 
-            Optional<User> user1 = dbExecutorOrm.load(connection, id, User.class);
+            Optional<User> user1 = userOrm.load(connection, id, User.class);
             System.out.println(user1);
             Assert.assertEquals(user, user1.orElse(null));
-            Assert.assertEquals(25, user1.orElse(null).getAge());
+            Assert.assertEquals(25, user1.get().getAge());
 
             // Account
+            DbExecutorOrm<Account> accountOrm = new DbExecutorOrmImpl<>();
             Account account = new Account(null, "test", BigDecimal.valueOf(0));
-            id = (Long) dbExecutorOrm.create(connection, account);
+            id = (Long) accountOrm.create(connection, account);
             Assert.assertEquals(Long.valueOf(1L), id);
 
             account.setRest(BigDecimal.valueOf(25.00));
-            dbExecutorOrm.update(connection, account);
+            accountOrm.update(connection, account);
 
             connection.commit();
 
-            Optional<Account> account1 = dbExecutorOrm.load(connection, id, Account.class);
+            Optional<Account> account1 = accountOrm.load(connection, id, Account.class);
             System.out.println(account1);
             Assert.assertEquals(account, account1.orElse(null));
             Assert.assertEquals(true, BigDecimal.valueOf(25.00).compareTo(account1.get().getRest()) == 0);
 
             account.setAccountType("test33");
-            dbExecutorOrm.createOrUpdate(connection, account);
+            accountOrm.createOrUpdate(connection, account);
             final Account account2 = new Account(null, "test", BigDecimal.valueOf(500));
-            dbExecutorOrm.createOrUpdate(connection, account2);
+            accountOrm.createOrUpdate(connection, account2);
             Assert.assertNotNull(account2.getNo());
         }
     }
